==> notebooks/03_define_files_dataset.ipynb <==
# DEV

import pandas as pd


dev = pd.read_csv('../resources/reports/dev.csv')
dev_details = pd.read_csv('../resources/reports/dev_details.csv')

ecg = dev_details[dev_details['ecg']==True]

ecg.groupby('freq')['freq'].count()

len(ecg)

dev

df = pd.merge(dev,ecg,left_on='Filename',right_on='file').iloc[:,1:]


df['Seizure Type'].unique()

df.groupby('Seizure Type')['Seizure Type'].count()

df[df['Seizure Type'].isin(['FNSZ','GNSZ','CPSZ','TNSZ','TCSZ',])]

df.to_csv('../resources/reports/ecg_files_dataset_especific_types.csv',index=False)

# TRAIN

train = pd.read_csv('../resources/reports/train.csv')
train_details = pd.read_csv('../resources/reports/train_details.csv')

ecg_train = train_details[train_details['ecg']==True]
len(ecg_train)

df_train = pd.merge(train,ecg_train,left_on='Filename',right_on='file').iloc[:,1:]

df_train = df_train[df_train['Seizure Type'].isin(['FNSZ','GNSZ','CPSZ','TNSZ','TCSZ',])]

len(df_train)

ecg_train.groupby('freq')['freq'].count()

df_train.to_csv('../resources/reports/ecg_files_train.csv',index=False)

# Eval

eval = pd.read_csv('../resources/reports/eval.csv')
eval_details = pd.read_csv('../resources/reports/eval_details.csv')

ecg_eval = eval_details[eval_details['ecg']==True]
len(ecg_eval)

df_eval = pd.merge(eval,ecg_eval,left_on='Filename',right_on='file').iloc[:,1:]

df_eval = df_eval[df_eval['Seizure Type'].isin(['FNSZ','GNSZ','CPSZ','TNSZ','TCSZ',])]

len(df_eval)

ecg_eval.groupby('freq')['freq'].count()

df_eval.to_csv('../resources/reports/ecg_files_eval.csv',index=False)
