==> notebooks/07_build_dataset.ipynb <==
# Criando datasets

import pandas as pd
import numpy as np
import os
from os.path import join, dirname
from scipy.io import loadmat
from dotenv import load_dotenv
import glob


dotenv_path = join(dirname('__file__'), '.env')

load_dotenv(dotenv_path)

DATALAKE_PATH = os.environ.get("DATALAKE_PATH")
ROOT_PATH = os.environ.get("ROOT_PATH")
SOURCE_PATH = os.environ.get("SOURCE_PATH")

SUB_PATH = 'tuh_seizures/processed/features_v2'

def build_dataset(type_dataset:str):
    
    files = glob.glob(f'{DATALAKE_PATH}/{SUB_PATH}/{type_dataset}/*.mat')

    D = []

    for f in files:
        d = loadmat(f)

        data = {
            'entropy_log_energy': d['entropy_log_energy'][0],
            'entropy_shannon': d['entropy_shannon'][0],
            'kur_log_energy': d['kur_log_energy'][0],
            'kur_shannon': d['kur_shannon'][0],
            'kurs': d['kurs'][0],
            'skew_log_energy': d['skew_log_energy'][0],
            'skew_shannon': d['skew_shannon'][0],
            'skews': d['skews'][0],
            'var_log_energy': d['var_log_energy'][0],
            'var_shannon': d['var_shannon'][0],
            'vars': d['vars'][0],
            'label': np.repeat(d['label'][0],d['entropy_log_energy'][0].shape[0])
            }

        D.append(pd.DataFrame(data=data))

    df = pd.concat(D)

    df.to_parquet(f'../data/{type_dataset}_dataset.parquet',engine='pyarrow')
    

build_dataset('train')
